+++ src/ethereum_clis/transition_tool.py
@@ -19,6 +19,7 @@ from requests.exceptions import ReadTimeout
 from requests_unixsocket import Session  # type: ignore
 
 from ethereum_test_base_types import BlobSchedule
+from ethereum_test_base_types.composite_types import ForkBlobSchedule
 from ethereum_test_exceptions import ExceptionMapper
 from ethereum_test_forks import Fork
 from ethereum_test_forks.helpers import get_development_forks, get_forks
@@ -29,6 +30,7 @@ from .cli_types import (
     Traces,
     TransactionReceipt,
     TransactionTraces,
+    TransitionToolCLIInput,
     TransitionToolContext,
     TransitionToolInput,
     TransitionToolOutput,
@@ -75,6 +77,7 @@ class TransitionTool(EthereumCLI):
     supports_opcode_count: ClassVar[bool] = False
 
     supports_xdist: ClassVar[bool] = True
+    supports_blob_params: ClassVar[bool] = False
 
     @abstractmethod
     def __init__(
@@ -175,6 +178,16 @@ class TransitionTool(EthereumCLI):
                 timestamp=self.env.timestamp,
             )
 
+        @property
+        def blob_params(self) -> ForkBlobSchedule | None:
+            """Return the blob parameters for the current fork."""
+            if self.blob_schedule:
+                fork_name = self.fork.fork_at(
+                    block_number=self.env.number, timestamp=self.env.timestamp
+                ).name()
+                return self.blob_schedule[fork_name]
+            return None
+
         def __post_init__(self) -> None:
             """Modify the reward if the environment number is 0."""
             if self.env.number == 0:
@@ -186,6 +199,15 @@ class TransitionTool(EthereumCLI):
                 alloc=self.alloc,
                 txs=self.txs,
                 env=self.env,
+            )
+
+        def to_cli_input(self) -> TransitionToolCLIInput:
+            """Convert the data to a TransitionToolCLIInput object."""
+            return TransitionToolCLIInput(
+                alloc=self.alloc,
+                txs=self.txs,
+                env=self.env,
+                blob_params=self.blob_params,
             )
 
         def get_request_data(self) -> TransitionToolRequest:
@@ -214,7 +236,7 @@ class TransitionTool(EthereumCLI):
         os.mkdir(os.path.join(temp_dir.name, "input"))
         os.mkdir(os.path.join(temp_dir.name, "output"))
 
-        input_contents = t8n_data.to_input().model_dump(mode="json", **model_dump_config)
+        input_contents = t8n_data.to_cli_input().model_dump(mode="json", **model_dump_config)
 
         input_paths = {
             k: os.path.join(temp_dir.name, "input", f"{k}.json") for k in input_contents.keys()
@@ -256,6 +278,13 @@ class TransitionTool(EthereumCLI):
                 [
                     "--opcode.count",
                     "opcodes.json",
+                ]
+            )
+        if self.supports_blob_params and input_paths.get("blobParams"):
+            args.extend(
+                [
+                    "--input.blobParams",
+                    input_paths["blobParams"],
                 ]
             )